Moved AddressContent and OrderAddress structs to Content/AddressContent.swift
